requirements.txt
fastapi
uvicorn
openai
pillow
requests
python-multipart
